dom/events/test/pointerevents/pointerlock/test_pointerevent_pointerlock_supercedes_capture-manual.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1399740
-->
  <head>
    <meta charset="utf-8">
    <title>Test for Bug 1399740</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <script src="/tests/SimpleTest/EventUtils.js"></script>
    <script type="text/javascript" src="mochitest_support_external.js"></script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
    <script type="text/javascript">
      SimpleTest.waitForExplicitFinish();

      function startTest() {
        runTestInNewWindow("pointerevent_pointerlock_supercedes_capture-manual.html");
      }

      function executeTest(int_win) {
        // Wait for the pointer is locked.
        int_win.document.addEventListener("pointerlockchange", () => {
          ok(int_win.document.pointerLockElement, "document must has been pointerlocked");
          int_win.document.addEventListener("pointerlockchange", () => {
              // Wait for the pointer is unlocked.
              sendMouseEvent(int_win, "div1", "mousemove", {inputSource:MouseEvent.MOZ_SOURCE_MOUSE});
              sendMouseEvent(int_win, "div1", "mousemove", {inputSource:MouseEvent.MOZ_SOURCE_MOUSE});
              sendMouseEvent(int_win, "div1", "mouseup", {inputSource:MouseEvent.MOZ_SOURCE_MOUSE});
            }, { once: true });
          sendMouseEvent(int_win, "div1", "mousemove", {inputSource:MouseEvent.MOZ_SOURCE_MOUSE});
          sendMouseEvent(int_win, "div1", "mousemove", {inputSource:MouseEvent.MOZ_SOURCE_MOUSE});
        }, { once: true });

        sendMouseEvent(int_win, "div1", "mousedown", {inputSource:MouseEvent.MOZ_SOURCE_MOUSE});
      }
    </script>
  </head>
  <body>
  </body>
</html>
